push files

+++ db_management/delete_product.php
@@ -1,9 +1,14 @@
 <?php
-echo print_r($_POST);
-require('connect.php');
-foreach ($_POST['product'] as $product) {
-	$sql = "DELETE FROM PRODUCTS WHERE name='$product'";
-	mysqli_query($conn, $sql);
+if ($_POST['submit'] == 'Delete') {
+	require('connect.php');
+	foreach ($_POST['product'] as $product) {
+		$sql = "DELETE FROM PRODUCTS WHERE name='$product'";
+		mysqli_query($conn, $sql);
+		header("Location:../create_prod.php");
+		die();
+	}
 }
-mysqli_close($conn);
-header("Location:../remove_product.php");
+
+if ($_POST['submit'] == 'Modify') {
+	
+}
